pallet_tips::migrations

Module unreserve_deposits

A migration that unreserves all funds held in the context of this pallet. A migration that unreserves all deposit and unlocks all stake held in the context of this pallet.

Structs

UnreserveDeposits

A migration that unreserves all tip deposits.

Traits

UnlockConfig

The configuration for UnreserveDeposits.