Trait pallet_tips::migrations::unreserve_deposits::UnlockConfig

pub trait UnlockConfig<I>: 'static {
    type Hash: Parameter;
    type AccountId: Parameter + Ord;
    type Currency: LockableCurrency<Self::AccountId> + ReservableCurrency<Self::AccountId>;
    type TipReportDepositBase: Get<<<Self as UnlockConfig<I>>::Currency as Currency<<Self as UnlockConfig<I>>::AccountId>>::Balance>;
    type DataDepositPerByte: Get<<<Self as UnlockConfig<I>>::Currency as Currency<<Self as UnlockConfig<I>>::AccountId>>::Balance>;
    type PalletName: Get<&'static str>;
    type DbWeight: Get<RuntimeDbWeight>;
    type BlockNumber: Parameter + Zero + Copy + Ord;
}

The configuration for UnreserveDeposits.

Required Associated Types

type Hash: Parameter

The hash used in the runtime.

type AccountId: Parameter + Ord

The account ID used in the runtime.

type Currency: LockableCurrency<Self::AccountId> + ReservableCurrency<Self::AccountId>

The currency type used in the runtime.

Should match the currency type previously used for the pallet, if applicable.

type TipReportDepositBase: Get<<<Self as UnlockConfig<I>>::Currency as Currency<<Self as UnlockConfig<I>>::AccountId>>::Balance>

Base deposit to report a tip.

Should match the currency type previously used for the pallet, if applicable.

type DataDepositPerByte: Get<<<Self as UnlockConfig<I>>::Currency as Currency<<Self as UnlockConfig<I>>::AccountId>>::Balance>

Deposit per byte to report a tip.

Should match the currency type previously used for the pallet, if applicable.

type PalletName: Get<&'static str>

The name of the pallet as previously configured in construct_runtime!.

type DbWeight: Get<RuntimeDbWeight>

The DB weight as configured in the runtime to calculate the correct weight.

type BlockNumber: Parameter + Zero + Copy + Ord

The block number as configured in the runtime.