Module pallet_tips::migrations::v4

Version 4.

For backward compatibility reasons, pallet-tips uses Treasury for storage module prefix before calling this migration. After calling this migration, it will get replaced with own storage identifier.

Functions

migrate

Migrate the entire storage of this pallet to a new prefix.

post_migrate

Some checks for after migration. This can be linked to frame_support::traits::OnRuntimeUpgrade::post_upgrade for further testing.

pre_migrate

Some checks prior to migration. This can be linked to frame_support::traits::OnRuntimeUpgrade::pre_upgrade for further testing.