Function pallet_tips::migrations::v4::migrate

pub fn migrate<T: Config, P: GetStorageVersion + PalletInfoAccess, N: AsRef<str>>(
    old_pallet_name: N
) -> Weight

Migrate the entire storage of this pallet to a new prefix.

This new prefix must be the same as the one set in construct_runtime. For safety, use PalletInfo to get it, as: <Runtime as frame_system::Config>::PalletInfo::name::<TipsPallet>.

The migration will look into the storage version in order not to trigger a migration on an up to date storage. Thus the on chain storage version must be less than 4 in order to trigger the migration.