Module pallet_tips::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

close_tip

Close and payout a tip.

report_awesome

Report something reason that deserves a tip and claim any eventual the finder’s fee.

retract_tip

Retract a prior tip-report from report_awesome, and cancel the process of tipping.

slash_tip

Remove and slash an already-open tip.

tip

Declare a tip value for an already-open tip.

tip_new

Give a tip for something new; no finder’s fee will be taken.