pallet_tips::pallet::dispatchables

Function report_awesome

pub fn report_awesome<T: Config<I>, I: 'static>(
    reason: Vec<u8>,
    who: <<T as Config>::Lookup as StaticLookup>::Source,
)

Report something reason that deserves a tip and claim any eventual the finder’s fee.

The dispatch origin for this call must be Signed.

Payment: TipReportDepositBase will be reserved from the origin account, as well as DataDepositPerByte for each byte in reason.

reason: The reason for, or the thing that deserves, the tip; generally this will be a UTF-8-encoded URL.

who: The account which should be credited for the tip.

Emits NewTip if successful.

Complexity

O(R) where R length of reason.

encoding and hashing of ‘reason’

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::report_awesome.