pallet_tips::pallet::dispatchables

Function retract_tip

pub fn retract_tip<T: Config<I>, I: 'static>(hash: T::Hash)

Retract a prior tip-report from report_awesome, and cancel the process of tipping.

If successful, the original deposit will be unreserved.

The dispatch origin for this call must be Signed and the tip identified by hash must have been reported by the signing account through report_awesome (and not through tip_new).

hash: The identity of the open tip for which a tip value is declared. This is formed as the hash of the tuple of the original tip reason and the beneficiary account ID.

Emits TipRetracted if successful.

Complexity

O(1)

Depends on the length of T::Hash which is fixed.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::retract_tip.