Function pallet_tips::pallet::dispatchables::tip_new

pub fn tip_new<T: Config<I>, I: 'static>(
    reason: Vec<u8>,
    who: <<T as Config>::Lookup as StaticLookup>::Source,
    tip_value: BalanceOf<T, I>,
)

Give a tip for something new; no finder’s fee will be taken.

The dispatch origin for this call must be Signed and the signing account must be a member of the Tippers set.

reason: The reason for, or the thing that deserves, the tip; generally this will be a UTF-8-encoded URL.

who: The account which should be credited for the tip.

tip_value: The amount of tip that the sender would like to give. The median tip value of active tippers will be given to the who.

Emits NewTip if successful.

Complexity

O(R + T) where R length of reason, T is the number of tippers.

O(T): decoding Tipper vec of length T. T is charged as upper bound given by ContainsLengthBound. The actual cost depends on the implementation of T::Tippers.

O(R): hashing and encoding of reason of length R

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::tip_new.