Module pallet_tips::pallet::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

Reasons

Simple preimage lookup from the reason’s hash to the original data. Again, has an insecure enumerable hash since the key is guaranteed to be the result of a secure hash.

Tips

TipsMap that are not yet completed. Keyed by the hash of (reason, who) from the value. This has the insecure enumerable hash function since the key itself is already guaranteed to be a secure hash.