pallet_tips::pallet

Type Alias Tips

pub type Tips<T: Config<I>, I: 'static = ()> = StorageMap<_GeneratedPrefixForStorageTips<T, I>, Twox64Concat, T::Hash, OpenTip<T::AccountId, BalanceOf<T, I>, BlockNumberFor<T>, T::Hash>, OptionQuery>;

TipsMap that are not yet completed. Keyed by the hash of (reason, who) from the value. This has the insecure enumerable hash function since the key itself is already guaranteed to be a secure hash.

Storage type is StorageMap with key type T :: Hash and value type OpenTip < T :: AccountId, BalanceOf < T, I > , BlockNumberFor < T > , T :: Hash >.

Aliased Type

pub struct Tips<T: Config<I>, I: 'static = ()>(/* private fields */);