Trait pallet_tips::Config

pub trait Config: Config + Config {
    type MaximumReasonLength: Get<u32>;
    type DataDepositPerByte: Get<BalanceOf<Self>>;
    type Tippers: SortedMembers<Self::AccountId> + ContainsLengthBound;
    type TipCountdown: Get<Self::BlockNumber>;
    type TipFindersFee: Get<Percent>;
    type TipReportDepositBase: Get<BalanceOf<Self>>;
    type Event: From<Event<Self>> + Into<Self::Event>;
    type WeightInfo: WeightInfo;
}

Associated Types

type MaximumReasonLength: Get<u32>

Maximum acceptable reason length.

type DataDepositPerByte: Get<BalanceOf<Self>>

The amount held on deposit per byte within the tip report reason or bounty description.

type Tippers: SortedMembers<Self::AccountId> + ContainsLengthBound

Origin from which tippers must come.

ContainsLengthBound::max_len must be cost free (i.e. no storage read or heavy operation).

type TipCountdown: Get<Self::BlockNumber>

The period for which a tip remains open after is has achieved threshold tippers.

type TipFindersFee: Get<Percent>

The percent of the final tip which goes to the original reporter of the tip.

type TipReportDepositBase: Get<BalanceOf<Self>>

The amount held on deposit for placing a tip report.

type Event: From<Event<Self>> + Into<Self::Event>

The overarching event type.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

Loading content...