Struct pallet_tips::pallet::Pallet

pub struct Pallet<T, I = ()>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn report_awesome(
origin: OriginFor<T>,
reason: Vec<u8>,
who: <<T as Config>::Lookup as StaticLookup>::Source
) -> DispatchResult

Report something reason that deserves a tip and claim any eventual the finder’s fee.

The dispatch origin for this call must be Signed.

Payment: TipReportDepositBase will be reserved from the origin account, as well as DataDepositPerByte for each byte in reason.

reason: The reason for, or the thing that deserves, the tip; generally this will be a UTF-8-encoded URL.

who: The account which should be credited for the tip.

Emits NewTip if successful.

Complexity

O(R) where R length of reason.

encoding and hashing of ‘reason’

pub fn retract_tip(origin: OriginFor<T>, hash: T::Hash) -> DispatchResult

Retract a prior tip-report from report_awesome, and cancel the process of tipping.

If successful, the original deposit will be unreserved.

The dispatch origin for this call must be Signed and the tip identified by hash must have been reported by the signing account through report_awesome (and not through tip_new).

hash: The identity of the open tip for which a tip value is declared. This is formed as the hash of the tuple of the original tip reason and the beneficiary account ID.

Emits TipRetracted if successful.

Complexity

O(1)

Depends on the length of T::Hash which is fixed.

pub fn tip_new(
origin: OriginFor<T>,
reason: Vec<u8>,
who: <<T as Config>::Lookup as StaticLookup>::Source,
tip_value: BalanceOf<T, I>
) -> DispatchResult

Give a tip for something new; no finder’s fee will be taken.

The dispatch origin for this call must be Signed and the signing account must be a member of the Tippers set.

reason: The reason for, or the thing that deserves, the tip; generally this will be a UTF-8-encoded URL.

who: The account which should be credited for the tip.

tip_value: The amount of tip that the sender would like to give. The median tip value of active tippers will be given to the who.

Emits NewTip if successful.

Complexity

O(R + T) where R length of reason, T is the number of tippers.

O(T): decoding Tipper vec of length T. T is charged as upper bound given by ContainsLengthBound. The actual cost depends on the implementation of T::Tippers.

O(R): hashing and encoding of reason of length R

pub fn tip(
origin: OriginFor<T>,
hash: T::Hash,
tip_value: BalanceOf<T, I>
) -> DispatchResult

Declare a tip value for an already-open tip.

The dispatch origin for this call must be Signed and the signing account must be a member of the Tippers set.

hash: The identity of the open tip for which a tip value is declared. This is formed as the hash of the tuple of the hash of the original tip reason and the beneficiary account ID.

tip_value: The amount of tip that the sender would like to give. The median tip value of active tippers will be given to the who.

Emits TipClosing if the threshold of tippers has been reached and the countdown period has started.

Complexity

O(T) where T is the number of tippers. decoding Tipper vec of length T, insert tip and check closing, T is charged as upper bound given by ContainsLengthBound. The actual cost depends on the implementation of T::Tippers.

Actually weight could be lower as it depends on how many tips are in OpenTip but it is weighted as if almost full i.e of length T-1.

pub fn close_tip(origin: OriginFor<T>, hash: T::Hash) -> DispatchResult

Close and payout a tip.

The dispatch origin for this call must be Signed.

The tip identified by hash must have finished its countdown period.

hash: The identity of the open tip for which a tip value is declared. This is formed as the hash of the tuple of the original tip reason and the beneficiary account ID.

Complexity

: O(T) where T is the number of tippers. decoding Tipper vec of length T. T is charged as upper bound given by ContainsLengthBound. The actual cost depends on the implementation of T::Tippers.

pub fn slash_tip(origin: OriginFor<T>, hash: T::Hash) -> DispatchResult

Remove and slash an already-open tip.

May only be called from T::RejectOrigin.

As a result, the finder is slashed and the deposits are lost.

Emits TipSlashed if successful.

Complexity

O(1).

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn tips<KArg>(
k: KArg
) -> Option<OpenTip<T::AccountId, BalanceOf<T, I>, T::BlockNumber, T::Hash>>where
KArg: EncodeLike<T::Hash>,

TipsMap that are not yet completed. Keyed by the hash of (reason, who) from the value. This has the insecure enumerable hash function since the key itself is already guaranteed to be a secure hash.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn reasons<KArg>(k: KArg) -> Option<Vec<u8>>where
KArg: EncodeLike<T::Hash>,

Simple preimage lookup from the reason’s hash to the original data. Again, has an insecure enumerable hash since the key is guaranteed to be the result of a secure hash.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn account_id() -> T::AccountId

The account ID of the treasury pot.

This actually does computation. If you need to keep using it, then make sure you cache the value and only call this once.

pub fn migrate_retract_tip_for_tip_new(module: &[u8], item: &[u8])